Moved obj store test sync util files to test/object-store/util/sync/
