+++ test/object-store/geospatial.cpp
@@ -36,7 +36,7 @@
 #include <realm/util/any.hpp>
 
 #if REALM_ENABLE_AUTH_TESTS
-#include <util/flx_sync_harness.hpp>
+#include <util/sync/flx_sync_harness.hpp>
 #endif // REALM_ENABLE_AUTH_TESTS
 
 #include <catch2/catch_all.hpp>
