+++ test/object-store/sync/app.cpp
@@ -17,11 +17,11 @@
 ////////////////////////////////////////////////////////////////////////////
 
 #include <collection_fixtures.hpp>
-#include <util/sync_test_utils.hpp>
-#include <util/baas_admin_api.hpp>
 #include <util/event_loop.hpp>
 #include <util/test_utils.hpp>
 #include <util/test_file.hpp>
+#include <util/sync/baas_admin_api.hpp>
+#include <util/sync/sync_test_utils.hpp>
 
 #include <realm/object-store/impl/object_accessor_impl.hpp>
 #include <realm/object-store/property.hpp>
