+++ test/object-store/sync/client_reset.cpp
@@ -17,12 +17,12 @@
 ////////////////////////////////////////////////////////////////////////////
 
 #include <collection_fixtures.hpp>
-#include <util/sync_test_utils.hpp>
-#include <util/baas_admin_api.hpp>
 #include <util/event_loop.hpp>
 #include <util/index_helpers.hpp>
 #include <util/test_file.hpp>
 #include <util/test_utils.hpp>
+#include <util/sync/baas_admin_api.hpp>
+#include <util/sync/sync_test_utils.hpp>
 
 #include <realm/object-store/thread_safe_reference.hpp>
 #include <realm/object-store/util/scheduler.hpp>
