+++ test/object-store/sync/metadata.cpp
@@ -16,9 +16,9 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#include <util/sync_test_utils.hpp>
 #include <util/test_path.hpp>
 #include <util/test_utils.hpp>
+#include <util/sync/sync_test_utils.hpp>
 
 #include <realm/object-store/object_schema.hpp>
 #include <realm/object-store/property.hpp>
